Split up recurrent_network_test
