+++ src/mlpack/tests/CMakeLists.txt
@@ -150,6 +150,7 @@ add_executable(mlpack_catch_test
   randomized_svd_test.cpp
   rbm_network_test.cpp
   recurrent_network_test.cpp
+  recurrent_network_2_test.cpp
   regularized_svd_test.cpp
   scaling_test.cpp
   serialization_catch.cpp
